requirements.txt
python-telegram-bot==21.0.1
torch
torchvision
Pillow